+++ api/add.php
@@ -14,6 +14,6 @@
 	echo api_add(isset($_GET['token']) 	 ? $_GET['token'] 	 : null,
 				 isset($_GET['shorturl']) ? $_GET['shorturl'] : null,
 				 isset($_GET['longurl'])  ? $_GET['longurl']  : null,
-				 isset($_GET['privateurl']) ? $_GET['privateurl'] : false);
+				 isset($_GET['unlistedurl']) ? $_GET['unlistedurl'] : false);
 
 ?>