+++ api/functions/func_modify.php
@@ -6,9 +6,9 @@
 	include_once($DOC_ROOT."/includes/logger.php");
 	include_once("func_authenticate.php");
 
-	function api_modify($token, $shorturl, $newlongurl, $privateurl){
+	function api_modify($token, $shorturl, $newlongurl, $unlistedurl){
 		global $ACTION_MOD_URL;
-		$privateurl = base_eng_2_int($privateurl);
+		$unlistedurl = base_eng_2_int($unlistedurl);
 		if(is_null($token)){
 			return json_encode(array('success' => false, 'reason' => 'Missing token'));
 		}
@@ -26,7 +26,7 @@
 		}
 
 		$conn = base_get_connection();
-		$sqlValidate1 = "SELECT short_url, long_url, private_url FROM translation WHERE short_url = '$shorturl'";
+		$sqlValidate1 = "SELECT short_url, long_url, unlisted_url FROM translation WHERE short_url = '$shorturl'";
 		$result = $conn->query($sqlValidate1);
 		$result = base_fetch_lazy($result);
 
@@ -35,16 +35,16 @@
 			return json_encode(array('success' => false, 'reason' => 'Unknown short URL'));
 		}
 
-		$sqlModify = "UPDATE translation SET long_url = '$newlongurl', private_url = '$privateurl' WHERE short_url = '$shorturl'";
+		$sqlModify = "UPDATE translation SET long_url = '$newlongurl', unlisted_url = '$unlistedurl' WHERE short_url = '$shorturl'";
 		$userid = cache_get_cached_user_id($token);
 		$oldlongurl = $result['long_url'];
 		$oldvalue = $shorturl." &rarr; ".$oldlongurl;
-		if($result['private_url'] == 1){
-			$oldvalue .= " (private)";
+		if($result['unlisted_url'] == 1){
+			$oldvalue .= " (unlisted)";
 		}
 		$newvalue = $shorturl." &rarr; ".$newlongurl;
-		if($privateurl == 1){
-			$newvalue .= " (private)";
+		if($unlistedurl == 1){
+			$newvalue .= " (unlisted)";
 		}
 		if(roles_is_admin(cache_get_cached_user($token))){
 			$result = $conn->query($sqlModify);