+++ api/modify.php
@@ -9,6 +9,6 @@
 	echo api_modify(isset($_GET['token']) ? $_GET['token'] : null,
 									isset($_GET['shorturl']) ? $_GET['shorturl'] : null,
 									isset($_GET['longurl']) ? $_GET['longurl'] : null,
-									isset($_GET['privateurl']) ? $_GET['privateurl'] : false);
+									isset($_GET['unlistedurl']) ? $_GET['unlistedurl'] : false);
 
 ?>