+++ js/api_com.js
@@ -15,11 +15,11 @@ function getToken() {
   }
   return null;
 }
-function setEditFields(short_url, long_url, private_url) {
+function setEditFields(short_url, long_url, unlisted_url) {
 	$("#ed_short_url").val(short_url);
 	$("#ed_long_url").val(long_url);
 	$("#ed_short_url_disabled").val(short_url);
-  $("#ed_private_url").prop("checked", private_url);
+  $("#ed_unlisted_url").prop("checked", unlisted_url);
 }
 function setConfirmFields(short_url) {
 	$("#conf_short_url").val(short_url);