+++ list.php
@@ -86,7 +86,7 @@
 								{ data: "dateadded" },
 								{ data: "views" },
 								{ data: "owner" },
-								{ data: "privateurl" },
+								{ data: "unlistedurl" },
 								{ data: null,
 									"target": 5,
 									"defaultContent": "<button class='btn btn-primary' data-toggle='modal' data-target='#edit_modal'><span class='fa fa-pencil'></span></button>&nbsp;&nbsp;&nbsp;<button class='btn btn-danger' data-toggle='modal' data-target='#confirm_modal'><span class='fa fa-trash-o'></span></button>"
@@ -96,7 +96,7 @@
 					}
 					$("#link_table").on('click','button',function(){
 						var data = table.row($(this).parents('tr')).data();
-						setEditFields(data['string'], data['longurl'], data['privateurl']);
+						setEditFields(data['string'], data['longurl'], data['unlistedurl']);
 						setConfirmFields(data['string']);
 					});
 				});